requirements.txt
flask
langgraph
torch
transformers
parler-tts
soundfile
python-dotenv
requests

